Move getEntityRecord type definition to types/selectors.ts, use via @typedef for now
