+++ packages/core-data/src/hooks/use-query-select.ts
@@ -1,7 +1,6 @@
 /**
  * WordPress dependencies
  */
-// @ts-ignore
 import { useSelect } from '@wordpress/data';
 
 /**
