--- packages/core-data/src/wordpress-data.d.ts
@@ -1,3 +1,0 @@
-declare module '@wordpress/data' {
-	export function createRegistrySelector< T extends (selector) => F, F extends Function >( fn: T, ...any ) : F;
-}
