--- packages/core-data/tsconfig.json
@@ -1,22 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"rootDir": "src",
-		"declarationDir": "build-types",
-		"noUnusedParameters": false,
-	},
-	"references": [
-		{ "path": "../api-fetch" },
-		{ "path": "../data" },
-		{ "path": "../deprecated" },
-		{ "path": "../element" },
-		{ "path": "../i18n" },
-		{ "path": "../is-shallow-equal" }
-	],
-	"include": [
-		"src/**/*.ts"
-	],
-	"exclude": [
-		"src/hooks/**/*.ts"
-	]
-}
